Added AuthMiddleware to check access token.
