+++ src/routes/diaryRouter.js
@@ -1,9 +1,9 @@
 import express from 'express';
 const diaryRouter = express.Router();
 import DiaryController from '../controllers/diaryController.js';
-import SecurityUtils from '../utils/securityUtils.js';
+import AuthMiddleware from '../routes/middleware/authMiddleware.js';
 
-diaryRouter.use(SecurityUtils.checkToken);
+diaryRouter.use(AuthMiddleware.checkToken);
 diaryRouter.use(express.json());
 
 diaryRouter
