+++ src/routes/markerMealRouter.js
@@ -1,9 +1,9 @@
 import express from 'express';
 const markerMealRouter = express.Router();
 import MarkerMealController from '../controllers/markerMealController.js';
-import SecurityUtils from '../utils/securityUtils.js';
+import AuthMiddleware from '../routes/middleware/authMiddleware.js';
 
-markerMealRouter.use(SecurityUtils.checkToken);
+markerMealRouter.use(AuthMiddleware.checkToken);
 markerMealRouter.use(express.json());
 
 markerMealRouter
