+++ src/routes/resetPasswordRouter.js
@@ -1,9 +1,9 @@
 import express from 'express';
 const resetPasswordRouter = express.Router();
 import ResetPasswordController from '../controllers/resetPasswordController.js';
-import SecurityUtils from '../utils/securityUtils.js';
+import AuthMiddleware from '../routes/middleware/authMiddleware.js';
 
-resetPasswordRouter.use(SecurityUtils.checkToken);
+resetPasswordRouter.use(AuthMiddleware.checkToken);
 resetPasswordRouter.use(express.json());
 
 resetPasswordRouter
